Trocado a tipagem dos valores de long para int, pois não era necessário 8 bytes para guardar os valores da matriz

+++ SmithSequencial.c
@@ -17,12 +17,12 @@
 #define ESQUERDA 1
 
 typedef struct smithWaterman {
-  long valor;
-  long origem;
+  int valor;
+  int origem;
 } smithWaterman;
 
 char *sequenciaA, *sequenciaB, *alinhamentoA, *alinhamentoB;
-long tamSequencia, maxI, maxJ, tamAlinOtimos;
+int tamSequencia, maxI, maxJ, tamAlinOtimos;
 smithWaterman *matrizValores;
 
 void alocarMatriz(){
@@ -47,7 +47,7 @@ void lerSequencias(char *nomeArq){
 	  exit(1);
 	}
 
-	fscanf(arq, "%ld", &tamSequencia);
+	fscanf(arq, "%d", &tamSequencia);
 
 	tamSequencia++;
 	
@@ -155,7 +155,7 @@ void imprimeMat(smithWaterman *mat){
 
   	for (int i = 0; i < tamSequencia; ++i){
   		for (int j = 0; j < tamSequencia; ++j){
-  			printf("%ld ", mat[i * tamSequencia + j].valor);
+  			printf("%d ", mat[i * tamSequencia + j].valor);
   		}
   		printf("\n");
   	}
@@ -166,7 +166,7 @@ void imprimeMat(smithWaterman *mat){
 
     for (int i = 0; i < tamSequencia; ++i){
     	for (int j = 0; j < tamSequencia; ++j){
-      		printf("%ld ", mat[i * tamSequencia + j].origem);
+      		printf("%d ", mat[i * tamSequencia + j].origem);
       	}
       printf("\n");
     }
@@ -176,7 +176,7 @@ void imprimeMat(smithWaterman *mat){
 int main(int argc, char **argv){
   
 	char *nomeArq;
- 	long maiorElemento;
+ 	int maiorElemento;
   	double start, end;
 
 	nomeArq = argv[1];
